+++ resources/views/nilai/table_mapel.blade.php
@@ -4,7 +4,9 @@
             <tr>
                 <th>No</th>
                 <th>Mata Pelajaran</th>
-                <th>Aksi</th>
+                @if (Auth::user()->isAdmin())
+                    <th>Aksi</th>
+                @endif
             </tr>
         </thead>
         <tbody>
@@ -12,9 +14,11 @@
                 <tr>
                     <td class="text-center align-middle">{{ $loop->iteration }}</td>
                     <td class="text-center align-middle">{{ $mapel->nama }}</td>
-                    <td class="text-center align-middle">
-                        <button class="btn btn-danger" data-url="{{ route('nilai.destroy_mapel', [$id_kelas, $id_tahun_ajar, $mapel->id]) }}" onclick="deleteMapelFromNilai(this)">Hapus</button>
-                    </td>
+                    @if (Auth::user()->isAdmin())
+                        <td class="text-center align-middle">
+                            <button class="btn btn-danger" data-url="{{ route('nilai.destroy_mapel', [$id_kelas, $id_tahun_ajar, $mapel->id]) }}" onclick="deleteMapelFromNilai(this)">Hapus</button>
+                        </td>
+                    @endif
                 </tr>                
             @endforeach
         </tbody>
